+++ changelog.md
@@ -1,4 +1,4 @@
-
+- [2020-09-17] Fix warning in guess_dimensions
 - [2020-09-17] Add argument to choose more finely the memory keyring before
   constructing load commands.
 - [2020-09-16] Fix tuple concatenation issue in loading data